requirements.txt
flask>=3.0
google-genai>=1.0.0
